restructure infrastructure folder
