+++ domains/certificates/Query.API/API.AppTests/CertificateIssuingTests.cs
@@ -3,8 +3,10 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using API.AppTests.Extensions;
-using API.AppTests.Infrastructure;
-using API.AppTests.Infrastructure.WriteToConsole;
+using API.AppTests.Infrastructure.Attributes;
+using API.AppTests.Infrastructure.CollectionDefinitions;
+using API.AppTests.Infrastructure.Factories;
+using API.AppTests.Infrastructure.Testcontainers;
 using API.AppTests.Mocks;
 using API.Query.API.ApiModels.Responses;
 using FluentAssertions;
