+++ domains/certificates/Query.API/API.AppTests/CertificatesByOwnerProjectionTest.cs
@@ -1,8 +1,9 @@
 using System;
 using System.Threading.Tasks;
 using AggregateRepositories;
-using API.AppTests.Infrastructure;
-using API.AppTests.Infrastructure.WriteToConsole;
+using API.AppTests.Infrastructure.Attributes;
+using API.AppTests.Infrastructure.CollectionDefinitions;
+using API.AppTests.Infrastructure.Testcontainers;
 using API.Query.API.Projections;
 using API.Query.API.Projections.Views;
 using CertificateEvents.Aggregates;
