+++ domains/certificates/Query.API/API.AppTests/Infrastructure/Attributes/TestPriorityAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace API.AppTests.Infrastructure.TestPriority;
+namespace API.AppTests.Infrastructure.Attributes;
 
 [AttributeUsage(AttributeTargets.Method)]
 public class TestPriorityAttribute : Attribute
