+++ domains/certificates/Query.API/API.AppTests/Infrastructure/CollectionDefinitions/StartupFixture.cs
@@ -3,7 +3,7 @@ using Serilog;
 using Serilog.Extensions.Logging;
 using Xunit;
 
-namespace API.AppTests.Infrastructure;
+namespace API.AppTests.Infrastructure.CollectionDefinitions;
 
 public class StartupFixture
 {
