+++ domains/certificates/Query.API/API.AppTests/Infrastructure/Testcontainers/MartenDbContainer.cs
@@ -4,7 +4,7 @@ using DotNet.Testcontainers.Configurations;
 using DotNet.Testcontainers.Containers;
 using Xunit;
 
-namespace API.AppTests.Infrastructure;
+namespace API.AppTests.Infrastructure.Testcontainers;
 
 public class MartenDbContainer : IAsyncLifetime
 {
