+++ domains/certificates/Query.API/API.AppTests/SwaggerTests.cs
@@ -1,7 +1,8 @@
 using System.Net;
 using System.Threading.Tasks;
-using API.AppTests.Infrastructure;
-using API.AppTests.Infrastructure.WriteToConsole;
+using API.AppTests.Infrastructure.Attributes;
+using API.AppTests.Infrastructure.CollectionDefinitions;
+using API.AppTests.Infrastructure.Factories;
 using FluentAssertions;
 using Microsoft.AspNetCore.Hosting;
 using VerifyXunit;
