+++ domains/certificates/Query.API/API.AppTests/TransferTests.cs
@@ -6,8 +6,10 @@ using System.Net.Http.Json;
 using System.Threading.Tasks;
 using API.AppTests.Extensions;
 using API.AppTests.Helpers;
-using API.AppTests.Infrastructure;
-using API.AppTests.Infrastructure.WriteToConsole;
+using API.AppTests.Infrastructure.Attributes;
+using API.AppTests.Infrastructure.CollectionDefinitions;
+using API.AppTests.Infrastructure.Factories;
+using API.AppTests.Infrastructure.Testcontainers;
 using API.AppTests.Mocks;
 using API.Query.API.ApiModels.Responses;
 using FluentAssertions;
